Adding nodes to buckets is functional
Changed names of old bucket functionality to make it not clash with our own functions.

Added bucket struct.

Changed function NewKademliaID to instead accept bytearray to make it more in line with our messaging structure.

Added addToBucket function which searches for an ID in a bucket and if it finds the ID it moves the entry to front, otherwise creates a new node instance to add to bucket.